+++ src/lib.rs
@@ -23,9 +23,9 @@
 //! ## Example
 //! ```
 //! use miniconf::{Miniconf, MiniconfAtomic};
-//! use serde::Deserialize;
+//! use serde::{Serialize, Deserialize};
 //!
-//! #[derive(Deserialize, MiniconfAtomic, Default)]
+//! #[derive(Deserialize, Serialize, MiniconfAtomic, Default)]
 //! struct Coefficients {
 //!     forward: f32,
 //!     backward: f32,
@@ -42,13 +42,13 @@
 //! let mut settings = Settings::default();
 //!
 //! // Update sample rate.
-//! miniconf::update(&mut settings, "sample_rate", b"350").unwrap();
+//! settings.set("sample_rate", b"350").unwrap();
 //!
 //! // Update filter coefficients.
-//! miniconf::update(&mut settings, "filter", b"{\"forward\": 35.6, \"backward\": 0.0}").unwrap();
+//! settings.set("filter", b"{\"forward\": 35.6, \"backward\": 0.0}").unwrap();
 //!
 //! // Update channel gain for channel 0.
-//! miniconf::update(&mut settings, "channel_gain/0", b"15").unwrap();
+//! settings.set("channel_gain/0", b"15").unwrap();
 //! ```
 //!
 //! ## Limitations
@@ -60,6 +60,8 @@
 #[cfg(feature = "mqtt-client")]
 mod mqtt_client;
 
+pub mod iter;
+
 #[cfg(feature = "mqtt-client")]
 pub use mqtt_client::MqttClient;
 
@@ -70,13 +72,18 @@ pub use minimq;
 pub use minimq::embedded_time;
 
 #[doc(hidden)]
-pub use serde::de::{Deserialize, DeserializeOwned};
+pub use serde::{
+    de::{Deserialize, DeserializeOwned},
+    ser::Serialize,
+};
 
 pub use serde_json_core;
 
 pub use derive_miniconf::{Miniconf, MiniconfAtomic};
 
 pub use heapless;
+
+use core::fmt::Write;
 
 /// Errors that occur during settings configuration
 #[derive(Debug, PartialEq)]
@@ -107,10 +114,24 @@ pub enum Error {
     /// Check that the serialized data is valid JSON and of the correct type.
     Deserialization(serde_json_core::de::Error),
 
+    /// The value provided could not be serialized.
+    ///
+    /// Check that the buffer had sufficient space.
+    SerializationFailed,
+
     /// When indexing into an array, the index provided was out of bounds.
     ///
     /// Check array indices to ensure that bounds for all paths are respected.
     BadIndex,
+}
+
+#[derive(Debug)]
+pub enum IterError {
+    /// The provided state vector is not long enough.
+    InsufficientStateDepth,
+
+    /// The provided topic length is not long enough.
+    InsufficientTopicLength,
 }
 
 impl From<Error> for u8 {
@@ -122,6 +143,7 @@ impl From<Error> for u8 {
             Error::AtomicUpdateRequired => 4,
             Error::Deserialization(_) => 5,
             Error::BadIndex => 6,
+            Error::SerializationFailed => 7,
         }
     }
 }
@@ -132,36 +154,116 @@ impl From<serde_json_core::de::Error> for Error {
     }
 }
 
+/// Metadata about a settings structure.
+pub struct MiniconfMetadata {
+    /// The maximum length of a topic in the structure.
+    pub max_topic_size: usize,
+
+    /// The maximum recursive depth of the structure.
+    pub max_depth: usize,
+}
+
 pub trait Miniconf {
+    /// Update settings directly from a string path and data.
+    ///
+    /// # Args
+    /// * `path` - The path to update within `settings`.
+    /// * `data` - The serialized data making up the contents of the configured value.
+    ///
+    /// # Returns
+    /// The result of the configuration operation.
+    fn set(&mut self, path: &str, data: &[u8]) -> Result<(), Error> {
+        self.string_set(path.split('/').peekable(), data)
+    }
+
+    /// Retrieve a serialized settings value from a string path.
+    ///
+    /// # Args
+    /// * `path` - The path to retrieve.
+    /// * `data` - The location to serialize the data into.
+    ///
+    /// # Returns
+    /// The number of bytes used in the `data` buffer for serialization.
+    fn get(&self, path: &str, data: &mut [u8]) -> Result<usize, Error> {
+        self.string_get(path.split('/').peekable(), data)
+    }
+
+    /// Create an iterator to read all possible settings paths.
+    ///
+    /// # Note
+    /// The state vector can be used to resume iteration from a previous point in time. The data
+    /// should be zero-initialized if starting iteration for the first time.
+    ///
+    /// # Template Arguments
+    /// * `TS` - The maximum number of bytes to encode a settings path into.
+    ///
+    /// # Args
+    /// * `state` - A state vector to record iteration state in.
+    fn into_iter<'a, const TS: usize>(
+        &'a self,
+        state: &'a mut [usize],
+    ) -> Result<iter::MiniconfIter<'a, Self, TS>, IterError> {
+        let metadata = self.get_metadata();
+
+        if TS < metadata.max_topic_size {
+            return Err(IterError::InsufficientTopicLength);
+        }
+
+        if state.len() < metadata.max_depth {
+            return Err(IterError::InsufficientStateDepth);
+        }
+
+        Ok(iter::MiniconfIter {
+            settings: self,
+            state,
+        })
+    }
+
+    /// Create an iterator to read all possible settings paths.
+    ///
+    /// # Note
+    /// This does not check that the topic size or state vector are large enough. If they are not,
+    /// panics may be generated internally by the library.
+    ///
+    /// # Note
+    /// The state vector can be used to resume iteration from a previous point in time. The data
+    /// should be zero-initialized if starting iteration for the first time.
+    ///
+    /// # Template Arguments
+    /// * `TS` - The maximum number of bytes to encode a settings path into.
+    ///
+    /// # Args
+    /// * `state` - A state vector to record iteration state in.
+    fn unchecked_into_iter<'a, const TS: usize>(
+        &'a self,
+        state: &'a mut [usize],
+    ) -> iter::MiniconfIter<'a, Self, TS> {
+        iter::MiniconfIter {
+            settings: self,
+            state,
+        }
+    }
+
     fn string_set(
         &mut self,
         topic_parts: core::iter::Peekable<core::str::Split<char>>,
         value: &[u8],
     ) -> Result<(), Error>;
 
-    fn recursive_iter<const TS: usize, const VS: usize>(
+    fn string_get(
+        &self,
+        topic_parts: core::iter::Peekable<core::str::Split<char>>,
+        value: &mut [u8],
+    ) -> Result<usize, Error>;
+
+    /// Get metadata about the settings structure.
+    fn get_metadata(&self) -> MiniconfMetadata;
+
+    fn recurse_paths<const TS: usize>(
         &self,
         index: &mut [usize],
-        _topic: &mut heapless::String<TS>,
-        value: &mut heapless::String<VS>,
+        topic: &mut heapless::String<TS>,
     ) -> Option<()>;
-}
-
-/// Convenience function to update settings directly from a string path and data.
-///
-/// # Note
-/// When using prefixes on the path, it is often simpler to call
-/// `Settings::string_set(path.peekable(), data)` directly.
-///
-/// # Args
-/// * `settings` - The settings to update
-/// * `path` - The path to update within `settings`.
-/// * `data` - The serialized data making up the contents of the configured value.
-///
-/// # Returns
-/// The result of the configuration operation.
-pub fn update<T: Miniconf>(settings: &mut T, path: &str, data: &[u8]) -> Result<(), Error> {
-    settings.string_set(path.split('/').peekable(), data)
 }
 
 macro_rules! impl_single {
@@ -179,33 +281,49 @@ macro_rules! impl_single {
                 Ok(())
             }
 
+            fn string_get(
+                &self,
+                mut topic_parts: core::iter::Peekable<core::str::Split<char>>,
+                value: &mut [u8],
+            ) -> Result<usize, Error> {
+                if topic_parts.peek().is_some() {
+                    return Err(Error::PathTooLong);
+                }
+
+                serde_json_core::to_slice(self, value).map_err(|_| Error::SerializationFailed)
+            }
+
+            fn get_metadata(&self) -> MiniconfMetadata {
+                MiniconfMetadata {
+                    // No topic length is needed, as there are no sub-members.
+                    max_topic_size: 0,
+                    // One index is required for the current element.
+                    max_depth: 1,
+                }
+            }
+
             // This implementation is the base case for primitives where it will
             // yield once for self, then return None on subsequent calls.
-            fn recursive_iter<const TS: usize, const VS: usize>(
+            fn recurse_paths<const TS: usize>(
                 &self,
                 index: &mut [usize],
                 _topic: &mut heapless::String<TS>,
-                value: &mut heapless::String<VS>,
-            ) -> Option<()>
-            where
-                Self: serde::Serialize,
-            {
+            ) -> Option<()> {
                 if index.len() == 0 {
-                    // I don't expect this to happen...
-                    panic!("index stack too small");
-                    // return None;
+                    // Note: During expected execution paths using `into_iter()`, the size of the
+                    // index stack is checked in advance to make sure this condition doesn't occur.
+                    // However, it's possible to happen if the user manually calls `recurse_paths`.
+                    unreachable!("Index stack too small");
                 }
 
                 let i = index[0];
                 index[0] += 1;
                 index[1..].iter_mut().for_each(|x| *x = 0);
 
-                match i {
-                    0 => {
-                        *value = serde_json_core::to_string(&self).unwrap();
-                        Some(())
-                    }
-                    _ => None,
+                if i == 0 {
+                    Some(())
+                } else {
+                    None
                 }
             }
         }
@@ -237,13 +355,104 @@ impl<T: Miniconf, const N: usize> Miniconf for [T; N] {
         Ok(())
     }
 
-    fn recursive_iter<const TS: usize, const VS: usize>(
+    fn string_get(
         &self,
-        _index: &mut [usize],
-        _topic: &mut heapless::String<TS>,
-        _value: &mut heapless::String<VS>,
+        mut topic_parts: core::iter::Peekable<core::str::Split<char>>,
+        value: &mut [u8],
+    ) -> Result<usize, Error> {
+        let next = topic_parts.next();
+        if next.is_none() {
+            return Err(Error::PathTooShort);
+        }
+
+        // Parse what should be the index value
+        let i: usize = serde_json_core::from_str(next.unwrap())
+            .or(Err(Error::BadIndex))?
+            .0;
+
+        if i >= self.len() {
+            return Err(Error::BadIndex);
+        }
+
+        self[i].string_get(topic_parts, value)
+    }
+
+    fn get_metadata(&self) -> MiniconfMetadata {
+        // First, figure out how many digits the maximum index requires when printing.
+        let mut index = N - 1;
+        let mut num_digits = 0;
+
+        while index > 0 {
+            index /= 10;
+            num_digits += 1;
+        }
+
+        let metadata = self[0].get_metadata();
+
+        // If the sub-members have topic size, we also need to include an additional character for
+        // the path separator. This is ommitted if the sub-members have no topic (e.g. fundamental
+        // types, enums).
+        if metadata.max_topic_size > 0 {
+            MiniconfMetadata {
+                max_topic_size: metadata.max_topic_size + num_digits + 1,
+                max_depth: metadata.max_depth + 1,
+            }
+        } else {
+            MiniconfMetadata {
+                max_topic_size: num_digits,
+                max_depth: metadata.max_depth + 1,
+            }
+        }
+    }
+
+    fn recurse_paths<const TS: usize>(
+        &self,
+        index: &mut [usize],
+        topic: &mut heapless::String<TS>,
     ) -> Option<()> {
-        unimplemented!("iteration not implemented for arrays yet")
+        let original_length = topic.len();
+
+        if index.len() == 0 {
+            // Note: During expected execution paths using `into_iter()`, the size of the
+            // index stack is checked in advance to make sure this condition doesn't occur.
+            // However, it's possible to happen if the user manually calls `recurse_paths`.
+            unreachable!("Index stack too small");
+        }
+
+        while index[0] < N {
+            // Add the array index to the topic name.
+            if topic.len() > 0 {
+                if topic.push('/').is_err() {
+                    // Note: During expected execution paths using `into_iter()`, the size of the
+                    // topic buffer is checked in advance to make sure this condition doesn't occur.
+                    // However, it's possible to happen if the user manually calls `recurse_paths`.
+                    unreachable!("Topic buffer too short");
+                }
+            }
+
+            if write!(topic, "{}", index[0]).is_err() {
+                // Note: During expected execution paths using `into_iter()`, the size of the
+                // topic buffer is checked in advance to make sure this condition doesn't occur.
+                // However, it's possible to happen if the user manually calls `recurse_paths`.
+                unreachable!("Topic buffer too short");
+            }
+
+            if self[index[0]]
+                .recurse_paths(&mut index[1..], topic)
+                .is_some()
+            {
+                return Some(());
+            }
+
+            // Strip off the previously prepended index, since we completed that element and need
+            // to instead check the next one.
+            topic.truncate(original_length);
+
+            index[0] += 1;
+            index[1..].iter_mut().for_each(|x| *x = 0);
+        }
+
+        None
     }
 }